Rename FieldsOverview to FieldList
