+++ graylog2-web-interface/src/views/components/sidebar/fields/FieldList.test.jsx
@@ -4,10 +4,10 @@ import { List } from 'immutable';
 
 import FieldType from 'views/logic/fieldtypes/FieldType';
 import FieldTypeMapping from 'views/logic/fieldtypes/FieldTypeMapping';
-import FieldsOverview from './FieldsOverview';
+import FieldList from './FieldList';
 
 
-describe('<FieldsOverview viewMetadata={viewMetadata} />', () => {
+describe('<FieldList viewMetadata={viewMetadata} />', () => {
   const properties = [{ enumerable: true }];
   const fieldType1 = new FieldType('string', properties, []);
   const fieldTypeMapping1 = new FieldTypeMapping('date', fieldType1);
@@ -23,13 +23,13 @@ describe('<FieldsOverview viewMetadata={viewMetadata} />', () => {
     id: 'aViewId',
   };
 
-  it('should render a FieldsOverview', () => {
-    const wrapper = mount(<FieldsOverview viewMetadata={viewMetadata} allFields={allFields} fields={fields} maximumHeight={1000} />);
+  it('should render a FieldList', () => {
+    const wrapper = mount(<FieldList viewMetadata={viewMetadata} allFields={allFields} fields={fields} maximumHeight={1000} />);
     expect(wrapper.find('span.field-element').text()).toBe('http_method');
   });
 
-  it('should render all fields in FieldsOverview after click', () => {
-    const wrapper = mount(<FieldsOverview viewMetadata={viewMetadata} allFields={allFields} fields={fields} maximumHeight={1000} />);
+  it('should render all fields in FieldList after click', () => {
+    const wrapper = mount(<FieldList viewMetadata={viewMetadata} allFields={allFields} fields={fields} maximumHeight={1000} />);
     expect(wrapper.find('span.field-element').length).toBe(1);
 
     wrapper.find('a[children="all"]').simulate('click');
@@ -44,7 +44,7 @@ describe('<FieldsOverview viewMetadata={viewMetadata} />', () => {
   });
 
   it('should search in the field list', () => {
-    const wrapper = mount(<FieldsOverview viewMetadata={viewMetadata} allFields={allFields} fields={fields} maximumHeight={1000} />);
+    const wrapper = mount(<FieldList viewMetadata={viewMetadata} allFields={allFields} fields={fields} maximumHeight={1000} />);
     expect(wrapper.find('span.field-element').length).toBe(1);
 
     wrapper.find('a[children="all"]').simulate('click');
@@ -59,7 +59,7 @@ describe('<FieldsOverview viewMetadata={viewMetadata} />', () => {
 
   it('should show hint when no fields are returned after filtering', () => {
     const hint = <i>No fields to show. Try changing your filter term or select a different field set above.</i>;
-    const wrapper = mount(<FieldsOverview viewMetadata={viewMetadata} allFields={allFields} fields={fields} maximumHeight={1000} />);
+    const wrapper = mount(<FieldList viewMetadata={viewMetadata} allFields={allFields} fields={fields} maximumHeight={1000} />);
     expect(wrapper).not.toContainReact(hint);
 
     wrapper.find('input#common-search-form-query-input').simulate('change', { target: { value: 'non_existing_field' } });
